Just some reorganization
