--- FinalEngine.Rendering/FaceCullMode.cs
@@ -1,13 +1,0 @@
-﻿// <copyright file="FaceCullMode.cs" company="Software Antics">
-//     Copyright (c) Software Antics. All rights reserved.
-// </copyright>
-
-namespace FinalEngine.Rendering
-{
-    public enum FaceCullMode
-    {
-        Front,
-
-        Back,
-    }
-}
