--- FinalEngine.Rendering/PrimitiveTopology.cs
@@ -1,17 +1,0 @@
-﻿// <copyright file="PrimitiveTopology.cs" company="Software Antics">
-//     Copyright (c) Software Antics. All rights reserved.
-// </copyright>
-
-namespace FinalEngine.Rendering
-{
-    public enum PrimitiveTopology
-    {
-        Line,
-
-        LineStrip,
-
-        Triangle,
-
-        TriangleStrip,
-    }
-}
