--- FinalEngine.Rendering/RasterMode.cs
@@ -1,13 +1,0 @@
-﻿// <copyright file="RasterMode.cs" company="Software Antics">
-//     Copyright (c) Software Antics. All rights reserved.
-// </copyright>
-
-namespace FinalEngine.Rendering
-{
-    public enum RasterMode
-    {
-        Solid,
-
-        Wireframe,
-    }
-}
