--- FinalEngine.Rendering/WindingDirection.cs
@@ -1,13 +1,0 @@
-﻿// <copyright file="WindingDirection.cs" company="Software Antics">
-//     Copyright (c) Software Antics. All rights reserved.
-// </copyright>
-
-namespace FinalEngine.Rendering
-{
-    public enum WindingDirection
-    {
-        Clockwise,
-
-        CounterClockwise,
-    }
-}
